refactor: add ability to get object properties
Create the PrintObjectProperties() method. This method allows for the retrieval of any object properties. This same logic can be used to simplify the menu class by creating a reusable menu. The new menu will print only the properties of the given object, allowing the user to specify the data in order to build a new object (of any inventory type)